<fix>[host]: fix enflame s60 not set to gpu device
Resolves: ZSTAC-75115
